+++ app/src/main/java/com/sysadmindoc/callshield/di/DetectionModule.kt
@@ -2,12 +2,15 @@ package com.sysadmindoc.callshield.di
 
 import com.sysadmindoc.callshield.data.CallbackDetector
 import com.sysadmindoc.callshield.data.CampaignDetector
+import com.sysadmindoc.callshield.data.CampaignObservationStore
 import com.sysadmindoc.callshield.data.HashWildcardMatcher
+import com.sysadmindoc.callshield.data.RoomCampaignObservationStore
 import com.sysadmindoc.callshield.data.SmsContentAnalyzer
 import com.sysadmindoc.callshield.data.SmsContextChecker
 import com.sysadmindoc.callshield.data.SpamHeuristics
 import com.sysadmindoc.callshield.data.SpamMLScorer
 import com.sysadmindoc.callshield.data.checker.CheckerDependencies
+import com.sysadmindoc.callshield.data.local.SpamDao
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -39,7 +42,11 @@ object DetectionModule {
 
     @Provides
     @Singleton
-    fun provideCampaignDetector(): CampaignDetector = CampaignDetector.shared
+    fun provideCampaignObservationStore(dao: SpamDao): CampaignObservationStore = RoomCampaignObservationStore(dao)
+
+    @Provides
+    @Singleton
+    fun provideCampaignDetector(store: CampaignObservationStore): CampaignDetector = CampaignDetector.shared.also { it.attachObservationStore(store) }
 
     @Provides
     @Singleton
@@ -47,5 +54,5 @@ object DetectionModule {
 
     @Provides
     @Singleton
-    fun provideCheckerDependencies(): CheckerDependencies = CheckerDependencies()
+    fun provideCheckerDependencies(campaignDetector: CampaignDetector): CheckerDependencies = CheckerDependencies(campaignDetector = campaignDetector)
 }